robot/requirements.txt
# Rovy Raspberry Pi Client Requirements
# Controls rover hardware and streams to PC server

# REST API Server (for mobile app)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# WebSocket client (for cloud streaming)
websockets>=11.0

# Serial communication with rover
pyserial>=3.5

# Camera
opencv-python>=4.5.0
# Note: depthai (OakD camera) is installed separately via system package manager
# For photo capture feature, ensure depthai is installed: pip install depthai

# Audio (ReSpeaker)
pyaudio>=0.2.11
numpy>=1.21.0

# For local audio playback
sounddevice>=0.4.0
soundfile>=0.12.0
scipy>=1.16.0  # For audio resampling

# Text-to-speech
piper-tts>=1.3.0

# HTTP client for gesture detection (async preferred, requests as fallback)
aiohttp>=3.8.0
requests>=2.28.0

# Wake word detection (Silero VAD + Whisper tiny)
silero-vad>=4.0.0
faster-whisper>=0.10.0  # Optimized Whisper for CPU inference
torch>=2.0.0  # Required for Silero VAD
torchaudio>=2.0.0  # Required for Silero VAD

# Systemd integration for watchdog notifications
systemd-python>=235

